LPS-66400 Do not retrieve the whole table to calculate if there are rows
There are drivers that have problems closing the resultSet if it is that huge

+++ modules/util/portal-tools-data-partitioning-sql-builder/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/exporter/BaseDataPartitioningExporter.java
@@ -174,11 +174,11 @@ public abstract class BaseDataPartitioningExporter
 		try (Connection connection = dataSource.getConnection();
 			PreparedStatement preparedStatement = buildPreparedStatement(
 				connection,
-				"select * from " + tableName +" where companyId = ?",
+				"select count(1) from " + tableName +" where companyId = ?",
 				companyId);
 			ResultSet resultSet = preparedStatement.executeQuery()) {
 
-			if (resultSet.next()) {
+			if (resultSet.next() && (resultSet.getInt(1) > 0)) {
 				String deleteSQL =
 					"delete from " + tableName + " where companyId = " +
 						companyId + ";\n";
